fix(packaging): harden Linux frame-copy delivery
