+++ package.json
@@ -212,7 +212,8 @@
         "tslib": "^2.8.1",
         "tsx": "4.21.0",
         "typescript": "5.9.3",
-        "typescript-eslint": "^8.46.2"
+        "typescript-eslint": "^8.46.2",
+        "yaml": "2.8.2"
     },
     "pnpm": {
         "overrides": {
