+++ tools/embedded-mpv/runtime-probe-contract.d.cts
@@ -1,0 +1,5 @@
+declare const RUNTIME_PROBE_CONTRACT: Readonly<{
+    readonly RUNTIME_PROBE_TIMEOUT_MS: 3000;
+}>;
+
+export = RUNTIME_PROBE_CONTRACT;
